requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.4.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
django-allauth==0.45.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-cors-headers==3.8.0
django-googledrive-storage==1.5.2
django==3.2.7
djangorestframework-simplejwt==5.0.0
djangorestframework==3.12.4
google-api-core==2.0.1
google-api-python-client==2.22.0
google-auth-httplib2==0.1.0
google-auth==2.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==3.2
oauth2client==4.1.3
oauthlib==3.1.1
pillow==8.3.2
pipfile==0.0.2
protobuf==3.18.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyjwt[crypto]==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.1
requests-oauthlib==1.3.0
requests==2.26.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.7
